requirements.txt
selenium==4.9.0
behave
webdriver-manager
pyyaml
allure-behave
webdriver-manager
pytest
pytest-xdist
autoit